Apply the same for wasi
